# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register-leader": {
      "post": {
        "operationId": "AuthController_registerLeader",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterLeaderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/teams": {
      "post": {
        "operationId": "AuthController_createTeam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOwnTeamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/campaigns": {
      "get": {
        "operationId": "CampaignsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Campaign"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      }
    },
    "/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      },
      "delete": {
        "operationId": "CampaignsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      }
    },
    "/activities": {
      "get": {
        "operationId": "ActivitiesController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actionDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activities"
        ]
      },
      "post": {
        "operationId": "ActivitiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activities"
        ]
      }
    },
    "/activities/{id}/availability": {
      "get": {
        "operationId": "ActivitiesController_availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actionDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-09-20",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityAvailabilityDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activities"
        ]
      }
    },
    "/activities/{id}": {
      "get": {
        "operationId": "ActivitiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activities"
        ]
      },
      "patch": {
        "operationId": "ActivitiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activities"
        ]
      },
      "delete": {
        "operationId": "ActivitiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "activities"
        ]
      }
    },
...
    "/dashboard/summary": {
      "get": {
        "operationId": "DashboardsController_summary",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dashboard"
        ]
      }
    },
    "/dashboard/by-activity": {
      "get": {
        "operationId": "DashboardsController_byActivity",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gincana Solidária API",
    "description": "Multi-tenant API for solidarity campaigns, submissions and scoring.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "member@gincana.local"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128,
            "example": "ChangeMe123!"
          },
          "deviceInfo": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterLeaderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160,
            "example": "Ana Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ana@exemplo.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          },
          "deviceInfo": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "CreateOwnTeamDto": {
        "type": "object",
        "properties": {
          "teamName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 160,
            "example": "Equipe Esperança"
          },
          "deviceInfo": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "teamName"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32
          },
          "deviceInfo": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "Campaign": {
        "type": "object",
        "properties": {}
      },
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 180
          },
          "description": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "CLOSED",
              "ARCHIVED"
            ]
          },
          "minimumActionsPerMonth": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "startsAt",
          "endsAt"
        ]
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 180
          },
          "description": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "status": {
            "enum": [
              "DRAFT",
              "ACTIVE",
              "CLOSED",
              "ARCHIVED"
            ],
            "type": "string"
          },
          "minimumActionsPerMonth": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "ActivityAvailabilityDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "nullable": true,
            "example": "Maximum reached"
          },
          "blockScope": {
            "nullable": true,
            "enum": [
              "CAMPAIGN",
              "MONTH",
              "DATE"
            ],
            "type": "string"
          },
          "blockedUntil": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "approvedOccurrences": {
            "type": "number",
            "minimum": 0
          },
          "approvedOccurrencesThisMonth": {
            "type": "number",
            "minimum": 0
          },
          "remainingOccurrences": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "remainingOccurrencesThisMonth": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          }
        },
        "required": [
          "available",
          "approvedOccurrences",
          "approvedOccurrencesThisMonth"
        ]
      },
      "Activity": {
        "type": "object",
        "properties": {}
      },
      "ActivityItemTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "casaco"
          },
          "pointsPerUnit": {
            "type": "number",
            "minimum": 0,
            "example": 25
          },
          "unit": {
            "type": "string",
            "maxLength": 40,
            "example": "unidade"
          },
          "minimumQuantity": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "pointsPerUnit",
          "unit"
        ]
      },
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 180
          },
          "description": {
            "type": "string"
          },
          "scoringType": {
            "enum": [
              "FIXED",
              "PER_ITEM",
              "PER_KG",
              "PER_MEMBER",
              "PER_COMPLETE_KIT",
              "TIERED",
              "MANUAL"
            ],
            "type": "string"
          },
          "points": {
            "type": "number",
            "minimum": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 40
          },
          "minimumQuantity": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "minimumParticipants": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "minimumParticipationPercent": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100
          },
          "maxOccurrences": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "maxOccurrencesPerMonth": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "maxOccurrencesPerParticipant": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "maxOccurrencesPerParticipantPerMonth": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "repeatable": {
            "type": "boolean",
            "default": true
          },
          "evidenceRequired": {
            "type": "boolean",
            "default": true
          },
          "rulesJson": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "itemTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityItemTypeDto"
            }
          }
        },
        "required": [
          "campaignId",
          "name",
          "scoringType"
        ]
      },
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 180
          },
          "description": {
            "type": "string"
          },
          "scoringType": {
            "enum": [
              "FIXED",
              "PER_ITEM",
              "PER_KG",
              "PER_MEMBER",
              "PER_COMPLETE_KIT",
              "TIERED",
              "MANUAL"
            ],
            "type": "string"
          },
          "points": {
            "type": "number",
            "minimum": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 40
          },
          "minimumQuantity": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "minimumParticipants": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "minimumParticipationPercent": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100
          },
          "maxOccurrences": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "maxOccurrencesPerMonth": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "maxOccurrencesPerParticipant": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "maxOccurrencesPerParticipantPerMonth": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "repeatable": {
            "type": "boolean",
            "default": true
          },
          "evidenceRequired": {
            "type": "boolean",
            "default": true
          },
          "rulesJson": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "itemTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityItemTypeDto"
            }
          }
        }
      },
...
    }
  }
}